refactor: removed `utils/general`, removed unused function `copyText`
