+++ components/Transaction/AddressTransactions.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image'
 import XECIcon from 'assets/xec-logo.png'
 import BCHIcon from 'assets/bch-logo.png'
 import EyeIcon from 'assets/eye-icon.png'
-import { FormatNumber } from 'utils/general'
+import { FormatNumber } from 'utils/index'
 import TableContainer from '../../components/TableContainer'
 import moment from 'moment'
 
@@ -51,7 +51,7 @@ export default ({ addressTransactions }: IProps): FunctionComponent => {
         Cell: (cellProps) => {
           const url = cellProps.cell.row.values.address.networkId === 1 ? 'https://explorer.e.cash/tx/' : 'https://blockchair.com/bitcoin-cash/transaction/'
           return (
-            <a href={url + cellProps.cell.value} target="_blank" rel="noopener noreferrer" className="table-eye-ctn">
+            <a href={url.concat(cellProps.cell.value)} target="_blank" rel="noopener noreferrer" className="table-eye-ctn">
               <div className="table-eye">
                 <Image src={EyeIcon} alt='View on explorer' />
               </div>
